+++ backend/models/Post.js
@@ -4,7 +4,8 @@ const getUserId = require('../helpers');
 class Post {
     static insert(req, callback) {
         const userId = getUserId(req);
-        connection.execute(`INSERT INTO Posts (user_id, title, image_url, content, published_date) values (?, ?, ?, ?, now());`, [userId, req.body.title, req.body.image_url, req.body.content], callback)
+        const imageUrl = `${req.protocol}://${req.get('host')}/media/${req.file.filename}`
+        connection.execute(`INSERT INTO Posts (user_id, title, image_url, content, published_date) values (?, ?, ?, ?, now());`, [userId, req.body.title, imageUrl, req.body.content], callback)
     };
 
     static update(req, callback) {
@@ -14,7 +15,17 @@ class Post {
     };
 
     static delete(req, callback) {
-        connection.execute(`DELETE FROM Posts WHERE id = ?`, [req.params.id], callback)
+        const userId = getUserId(req);
+        connection.execute(`SELECT posts.id as post_id, posts.content, users.id as user_id, users.is_admin FROM posts 
+        INNER JOIN users 
+        WHERE posts.id = ? AND users.id = ?`, [req.params.id, userId], (error, results) => {
+            const post = results[0];
+            // On vérifie si l'utilisateur est soit l'auteur du post soit un admin
+            if (post.is_admin === 1 || post.user_id === userId) {
+                // il peut supprimer
+                connection.execute(`DELETE FROM Posts WHERE id = ?`, [req.params.id], callback)
+            }
+        })
     };
 
     static getAll(callback) {
